rename costumer to customer and store historical data

+++ packages/common/dataplattform/common/handlers/process.py
@@ -83,10 +83,11 @@ class ProcessHandler:
         data = self.wrapped_func['process'](s3_data, event.get('Records', []))
         partitions = self.wrapped_func_args.get('process', {}).get('partitions', {})
         overwrite = self.wrapped_func_args.get('process', {}).get('overwrite', False)
+        historical_tables = self.wrapped_func_args.get('process', {}).get('historical_tables', [])
 
-        return data, partitions, overwrite
+        return data, partitions, overwrite, historical_tables
 
-    def update_tables(self, tables, partitions, overwrite):
+    def update_tables(self, tables, partitions, overwrite, historical_tables):
         updated_tables = []
 
         for table_name, frame in tables.items():
@@ -103,8 +104,7 @@ class ProcessHandler:
             frame = ensure_partitions_has_values(frame, table_partitions)
 
             table_exists = check_exists(self.s3, frame, table_name, table_partitions)
-
-            if table_exists and overwrite:
+            if table_exists and overwrite and (table_name not in historical_tables):
                 delete_table(self.s3, table_name)
                 table_exists = False
 
@@ -130,10 +130,11 @@ class ProcessHandler:
         sns.publish({'tables': updated_tables}, 'DataUpdate')
         return True
 
-    def process(self, partitions, overwrite=False):
+    def process(self, partitions, overwrite=False, historical_tables=[]):
         def wrap(f):
             self.wrapped_func['process'] = make_wrapper_func(f, dict)
-            self.wrapped_func_args['process'] = dict(partitions=partitions, overwrite=overwrite)
+            self.wrapped_func_args['process'] = dict(partitions=partitions, overwrite=overwrite,
+                                                     historical_tables=historical_tables)
             return self.wrapped_func['process']
 
         return wrap
@@ -145,7 +146,7 @@ class PersonDataProcessHandler(ProcessHandler):
         self.id_type = id_type
 
     def call_wrapped(self, s3_data, event):
-        data, partitions, overwrite = super().call_wrapped(s3_data, event)
+        data, partitions, overwrite, historical_tables = super().call_wrapped(s3_data, event)
 
         pdtables = self.wrapped_func_args.get('process', {}).get('person_data_tables', [])
 
@@ -161,15 +162,16 @@ class PersonDataProcessHandler(ProcessHandler):
                 indexNames = frame[frame['guid'] == ""].index
                 frame.drop(indexNames, inplace=True)
 
-        return data, partitions, overwrite
+        return data, partitions, overwrite, historical_tables
 
-    def process(self, partitions, person_data_tables: List[str], overwrite=False):
+    def process(self, partitions, person_data_tables: List[str], overwrite=False, historical_tables=[]):
         def wrap(f):
             self.wrapped_func['process'] = make_wrapper_func(f, dict)
             self.wrapped_func_args['process'] = dict(
                 partitions=partitions,
                 overwrite=overwrite,
-                person_data_tables=person_data_tables
+                person_data_tables=person_data_tables,
+                historical_tables=historical_tables
             )
             return self.wrapped_func['process']
